* pr-199: (44 commits)
  Always run python3 tests, they succeed now.
  ci: remove stretch tests
  test 2001-cruft: add cruft report
  test 2000-dinstall: run make-changelog
  Add test 0006-dm-commands
  Testsuite fixtures: change some uploaders and add extra package
  tests: 2003-c-s-propup: remove first dinstall call
  test 0005-encoding: add extra checks
  integration-tests: add check_output function
  run-ci: don't run 2to3, it's not needed anymore
  run-ci: add annotated coverage report
  gitlab-ci: coverage job only uses py2 data
  integration-tests/dak-coverage.sh: fix coverage when working dir changes
  integration-tests/common: add function to save ci artifacts
  run-ci: define DAK_CI_OUTPUT_DIR
  daklib/dbconn.py: use textutils.force_to_utf8 for contents filename
  daklib/utils.py: handle encoding in gpg_get_key_addresses using six
  daklib/checks.py: use six function in check_fields_for_valid_utf8
  daklib/dbconn.py: set database connection to utf-8 for python3
  load_keys: make sure data from gpg is str
  ...

+++ dak/import.py
@@ -139,7 +139,7 @@ def import_dump(log, transaction, suite, component, fh,
                 print('W: {0}: unknown fingerprint {1}'.format(filename, fpr))
 
         directory, filename = os.path.split(os.path.abspath(path))
-        hashed_file = daklib.upload.HashedFile(filename, long(size), md5, sha1, sha256)
+        hashed_file = daklib.upload.HashedFile(filename, int(size), md5, sha1, sha256)
         hashed_file.check(directory)
 
         import_file(log, transaction, suite, component, directory, hashed_file,
